changed InputAnalysisFilter thread frequency to a double as the thread functions expect

+++ source/dsp/InputAnalysisFilter.h
@@ -28,7 +28,7 @@ private:
     // Debug functions.
     void printBandMagnitudesPreProcessing();
 
-    static const uint16 ANALYSIS_FREQUENCY_MS;  //! How frequently the analysis is performed (in milliseconds).
+    static const double ANALYSIS_FREQUENCY_MS;  //! How frequently the analysis is performed (in milliseconds).
 
     typedef std::array< Filter, Equalizer::NUM_BANDS >                   SingleBandFilterSequence;
     typedef std::array< SingleBandFilterSequence, Equalizer::NUM_BANDS > FilterMatrix;
